pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sb3vm"
version = "0.1.0"
description = "Headless Scratch sb3 VM MVP"
requires-python = ">=3.10"

[project.optional-dependencies]
render = [
    "Pillow>=10.0.0",
    "cairosvg>=2.7.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
